Change all edits
