+++ Mukhin/lab3/lab3.sh
@@ -1,3 +1,3 @@
-g++ Source/lab3.cpp -o lab3
+g++ Source/main.cpp Source/lab3.hpp -o lab3
 ./lab3 ./Tests/in.txt ./Tests/out.txt
 rm lab3
